fix: ended at value and generic function
